requirements-minimal.txt
# WhisperType - Minimal requirements
faster-whisper>=1.0.0
pynput>=1.7.6
pyperclip>=1.8.2
PyQt5>=5.15.0
sounddevice>=0.4.6
audioplayer>=0.6
python-dotenv>=1.0.0
PyYAML>=6.0
webrtcvad-wheels>=2.0.11
numpy>=1.24.0
